+++ NEWS.md
@@ -31,3 +31,4 @@ v5 (2012-10-25)
 
 v6 (2012-11-05)
 - support for 3.7.1
+- bugfix: [terminal lost after screen is locked (github #6)](https://github.com/zzrough/gs-extensions-drop-down-terminal/issues/6)